Move RevocationOptions into crl module
